// src/types/clip.d.ts
export interface IClip {
  clipUID: string
  videoId: string
  clipTitle: string
  ownerChannelId: string
  ownerChannel: null
  thumbnailImageUrl: string
  categoryType: string
  clipCategory: string
  duration: number
  adult: boolean
  createdDate: string
  recId: string
  readCount: number
  blindType: null
  privateUserBlock: boolean
}